+++ src/app/layout/type/user-settings.type.ts
@@ -3,4 +3,6 @@ import { Language } from '@shared/module/poe/type';
 export interface UserSettings {
     leagueId?: string;
     language?: Language;
+    openUserSettingsKeybinding?: string;
+    exitAppKeybinding?: string;
 }