+++ src/app/shared/module/poe/service/stats-id/stats-id.service.spec.ts
@@ -1,8 +1,8 @@
-import { async, TestBed } from "@angular/core/testing";
+import { async, TestBed } from '@angular/core/testing';
 import { SharedModule } from '@shared/shared.module';
 import { StatsIdService } from './stats-id.service';
 
-fdescribe('StatsIdService', () => {
+describe('StatsIdService', () => {
     let sut: StatsIdService;
     beforeEach(async(() => {
         TestBed.configureTestingModule({